+++ apps/docs/src/content/04-components/structure/flex/examples/layoutCard.tsx
@@ -13,6 +13,6 @@ import {
     Quam.
   </Text>
   <Flex align="end" grow>
-    <Button color="success">Install</Button>
+    <Button>Install</Button>
   </Flex>
 </LayoutCard>;