+++ apps/docs/src/content/04-components/structure/list/examples/default.tsx
@@ -28,7 +28,7 @@ export default () => {
     >
       <DomainList.StaticData data={domains} />
       <ActionGroup>
-        <Button color="success">Anlegen</Button>
+        <Button>Anlegen</Button>
       </ActionGroup>
       <DomainList.Search />
       <DomainList.Filter