+++ packages/remote-elements/src/auto-generated/RemoteButtonElement.ts
@@ -45,7 +45,6 @@ export class RemoteButtonElement extends FlowRemoteElement<RemoteButtonElementPr
       lang: {},
       name: {},
       preventFocusOnPress: {},
-      render: {},
       size: {},
       slot: {},
       translate: {},