+++ packages/remote-elements/src/auto-generated/RemoteCheckboxButtonElement.ts
@@ -35,7 +35,6 @@ export class RemoteCheckboxButtonElement extends FlowRemoteElement<RemoteCheckbo
       isSelected: {},
       lang: {},
       name: {},
-      render: {},
       slot: {},
       translate: {},
       validate: {},