+++ packages/remote-elements/src/auto-generated/RemoteRadioElement.ts
@@ -24,7 +24,6 @@ export class RemoteRadioElement extends FlowRemoteElement<RemoteRadioElementProp
       inputRef: {},
       isDisabled: {},
       lang: {},
-      render: {},
       slot: {},
       translate: {},
       value: {},