+++ packages/remote-elements/src/auto-generated/RemoteRangeCalendarElement.ts
@@ -36,7 +36,6 @@ export class RemoteRangeCalendarElement extends FlowRemoteElement<RemoteRangeCal
       maxValue: {},
       minValue: {},
       pageBehavior: {},
-      render: {},
       selectionAlignment: {},
       slot: {},
       translate: {},